[backend] Move playbook notifier component into specific files
