+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-components.ts
@@ -16,8 +16,8 @@ import * as R from 'ramda';
 import { v4 as uuidv4 } from 'uuid';
 import type { JSONSchemaType } from 'ajv';
 import * as jsonpatch from 'fast-json-patch';
-import { type BasicStoreEntityPlaybook, ENTITY_TYPE_PLAYBOOK, type PlaybookComponent } from './playbook-types';
-import { AUTOMATION_MANAGER_USER, AUTOMATION_MANAGER_USER_UUID, executionContext, isUserCanAccessStixElement, isUserInPlatformOrganization, SYSTEM_USER } from '../../utils/access';
+import { type PlaybookComponent } from './playbook-types';
+import { AUTOMATION_MANAGER_USER, AUTOMATION_MANAGER_USER_UUID, executionContext, SYSTEM_USER } from '../../utils/access';
 import { pushToConnector, pushToWorkerForConnector } from '../../database/rabbitmq';
 import {
   ABSTRACT_STIX_CORE_OBJECT,
@@ -40,7 +40,7 @@ import type { BasicStoreCommon, BasicStoreRelation, StoreCommon, StoreRelation }
 import { generateInternalId, generateStandardId, idGenFromData } from '../../schema/identifier';
 import { now, observableValue, utcDate } from '../../utils/format';
 import type { StixCampaign, StixContainer, StixIncident, StixInfrastructure, StixMalware, StixReport, StixThreatActor } from '../../types/stix-2-1-sdo';
-import { convertStixToInternalTypes, generateInternalType, getParentTypes } from '../../schema/schemaUtils';
+import { generateInternalType, getParentTypes } from '../../schema/schemaUtils';
 import {
   ENTITY_TYPE_ATTACK_PATTERN,
   ENTITY_TYPE_CAMPAIGN,
@@ -55,9 +55,9 @@ import {
 import type { CyberObjectExtension, StixBundle, StixCoreObject, StixCyberObject, StixDomainObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
 import { STIX_EXT_MITRE, STIX_EXT_OCTI, STIX_EXT_OCTI_SCO } from '../../types/stix-2-1-extensions';
 import { connectorsForPlaybook } from '../../database/repository';
-import { internalFindByIds, fullEntitiesList, fullRelationsList, storeLoadById } from '../../database/middleware-loader';
+import { internalFindByIds, fullEntitiesList, fullRelationsList } from '../../database/middleware-loader';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../organization/organization-types';
-import { getEntitiesMapFromCache, getEntityFromCache } from '../../database/cache';
+import { getEntitiesMapFromCache } from '../../database/cache';
 import { createdBy, objectLabel, objectMarking } from '../../schema/stixRefRelationship';
 import { logApp } from '../../config/conf';
 import { FunctionalError } from '../../config/errors';
@@ -66,10 +66,6 @@ import { isEmptyField, isNotEmptyField, READ_RELATIONSHIPS_INDICES, READ_RELATIO
 import { schemaAttributesDefinition } from '../../schema/schema-attributes';
 import { schemaRelationsRefDefinition } from '../../schema/schema-relationsRef';
 import { stixLoadByIds } from '../../database/middleware';
-import { usableNotifiers } from '../notifier/notifier-domain';
-import { convertToNotificationUser, type DigestEvent, EVENT_NOTIFICATION_VERSION } from '../../manager/notificationManager';
-import { storeNotificationEvent } from '../../database/redis';
-import { ENTITY_TYPE_SETTINGS } from '../../schema/internalObject';
 import { isStixCyberObservable } from '../../schema/stixCyberObservable';
 import { createStixPattern } from '../../python/pythonBridge';
 import { generateKeyValueForIndicator } from '../../domain/stixCyberObservable';
@@ -85,18 +81,17 @@ import { ENTITY_TYPE_CONTAINER_TASK, type StixTask, type StoreEntityTask } from 
 import { EditOperation, FilterMode } from '../../generated/graphql';
 import { ENTITY_TYPE_MARKING_DEFINITION } from '../../schema/stixMetaObject';
 import { schemaTypesDefinition } from '../../schema/schema-types';
-import { generateCreateMessage } from '../../database/generate-message';
 import { ENTITY_TYPE_CONTAINER_CASE } from '../case/case-types';
 import { findAllByCaseTemplateId } from '../task/task-domain';
 import type { BasicStoreEntityTaskTemplate } from '../task/task-template/task-template-types';
-import type { BasicStoreSettings } from '../../types/settings';
 import { AUTHORIZED_MEMBERS_SUPPORTED_ENTITY_TYPES, editAuthorizedMembers } from '../../utils/authorizedMembers';
 import { removeOrganizationRestriction } from '../../domain/stix';
 import { ENTITY_TYPE_CONTAINER_GROUPING } from '../grouping/grouping-types';
 import { ENTITY_TYPE_CONTAINER_FEEDBACK } from '../case/feedback/feedback-types';
 import { PLAYBOOK_SEND_EMAIL_TEMPLATE_COMPONENT } from './components/send-email-template-component';
-import { convertMembersToUsers, extractBundleBaseElement } from './playbook-utils';
+import { extractBundleBaseElement } from './playbook-utils';
 import { PLAYBOOK_DATA_STREAM_PIR } from './playbookComponents/playbook-data-stream-pir-component';
+import { PLAYBOOK_NOTIFIER_COMPONENT } from './playbookComponents/playbook-notifier-component';
 
 // region built in playbook components
 interface LoggerConfiguration {
@@ -1188,74 +1183,6 @@ const PLAYBOOK_RULE_COMPONENT: PlaybookComponent<RuleConfiguration> = {
   }
 };
 
-export interface NotifierConfiguration {
-  notifiers: string[]
-  authorized_members: object
-}
-const PLAYBOOK_NOTIFIER_COMPONENT_SCHEMA: JSONSchemaType<NotifierConfiguration> = {
-  type: 'object',
-  properties: {
-    notifiers: {
-      type: 'array',
-      uniqueItems: true,
-      default: [],
-      $ref: 'Notifiers',
-      items: { type: 'string', oneOf: [] }
-    },
-    authorized_members: { type: 'object' },
-  },
-  required: ['notifiers', 'authorized_members'],
-};
-const PLAYBOOK_NOTIFIER_COMPONENT: PlaybookComponent<NotifierConfiguration> = {
-  id: 'PLAYBOOK_NOTIFIER_COMPONENT',
-  name: 'Send to notifier',
-  description: 'Send user notification',
-  icon: 'notification',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [],
-  configuration_schema: PLAYBOOK_NOTIFIER_COMPONENT_SCHEMA,
-  schema: async () => {
-    const context = executionContext('playbook_components');
-    const notifiers = await usableNotifiers(context, SYSTEM_USER);
-    const elements = notifiers.map((c) => ({ const: c.id, title: c.name }));
-    const schemaElement = { properties: { notifiers: { items: { oneOf: elements } } } };
-    return R.mergeDeepRight<JSONSchemaType<NotifierConfiguration>, any>(PLAYBOOK_NOTIFIER_COMPONENT_SCHEMA, schemaElement);
-  },
-  executor: async ({ dataInstanceId, playbookId, playbookNode, bundle }) => {
-    const context = executionContext('playbook_components');
-    const playbook = await storeLoadById<BasicStoreEntityPlaybook>(context, SYSTEM_USER, playbookId, ENTITY_TYPE_PLAYBOOK);
-    const { notifiers, authorized_members } = playbookNode.configuration;
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle);
-    const targetUsers = await convertMembersToUsers(authorized_members as { value: string }[], baseData, bundle);
-    const settings = await getEntityFromCache<BasicStoreSettings>(context, SYSTEM_USER, ENTITY_TYPE_SETTINGS);
-    const notificationsCall = [];
-    for (let index = 0; index < targetUsers.length; index += 1) {
-      const targetUser = targetUsers[index];
-      const user_inside_platform_organization = isUserInPlatformOrganization(targetUser, settings);
-      const userContext = { ...context, user_inside_platform_organization };
-      const stixElements = bundle.objects.filter((o) => isUserCanAccessStixElement(userContext, targetUser, o));
-      const notificationEvent: DigestEvent = {
-        version: EVENT_NOTIFICATION_VERSION,
-        playbook_source: playbook.name,
-        notification_id: playbookNode.id,
-        target: convertToNotificationUser(targetUser, notifiers),
-        type: 'digest',
-        data: stixElements.map((stixObject) => ({
-          notification_id: playbookNode.id,
-          instance: stixObject,
-          type: 'create', // TODO Improve that with type event follow up
-          message: generateCreateMessage({ ...stixObject, entity_type: convertStixToInternalTypes(stixObject.type) }) === '-' ? playbookNode.name : generateCreateMessage({ ...stixObject, entity_type: convertStixToInternalTypes(stixObject.type) }),
-        }))
-      };
-      notificationsCall.push(storeNotificationEvent(context, notificationEvent));
-    }
-    if (notificationsCall.length > 0) {
-      await Promise.all(notificationsCall);
-    }
-    return { output_port: undefined, bundle };
-  }
-};
 interface CreateIndicatorConfiguration {
   all: boolean
   wrap_in_container: boolean
